LengthOfLongestSubString
Commented solution is what I came up with on my own. NeatCode solution demonstrates that your solution can be O(n) even with a nested loop because the operations we are doing are different... the inner while loop does not iterate through the whole array, it simply iterates and slides the window.